Enemy Knight now has a tell before it throws the sword
For some reason, adding this feature meant that I had to separate the enemy throwing and player throwing logic into two separate scripts. For some reason, even though I didn't even touch it while adding this feature, the player throw logic broke. If i removed the "if not paused then fire" logic for the player, then it would fire. I don't know why this broke things, since it has nothing to do with eachother, and the logic worked with other objects, but it all works now and that's what matters.